Separating bilinear and nearest factories
