+++ ttnn/cpp/ttnn/operations/pool/CMakeLists.txt
@@ -36,7 +36,9 @@ target_sources(
         generic/generic_pools.cpp
         global_avg_pool/global_avg_pool.cpp
         grid_sample/device/grid_sample_op.cpp
-        grid_sample/device/grid_sample_program_factory.cpp
+        grid_sample/device/grid_sample_bilinear_program_factory.cpp
+        grid_sample/device/grid_sample_nearest_program_factory.cpp
+        grid_sample/device/grid_sample_utils.cpp
         grid_sample/grid_sample.cpp
         grid_sample/grid_sample_prepare_grid.cpp
         pool_utils.cpp
