+++ ttnn/cpp/ttnn/operations/pool/grid_sample/device/grid_sample_op.cpp
@@ -287,15 +287,25 @@ operation::ProgramWithCallbacks GridSample::create_program(
     const Tensor& grid_tensor = input_tensors.at(1);
     Tensor& output_tensor = output_tensors.at(0);
 
-    return grid_sample_program_factory(
-        input_tensor,
-        grid_tensor,
-        output_tensor,
-        mode_,
-        padding_mode_,
-        align_corners_,
-        use_precomputed_grid_,
-        batch_output_channels_);
+    if (mode_ == "bilinear") {
+        return grid_sample_bilinear_program_factory(
+            input_tensor,
+            grid_tensor,
+            output_tensor,
+            padding_mode_,
+            align_corners_,
+            use_precomputed_grid_,
+            batch_output_channels_);
+    } else {
+        return grid_sample_nearest_program_factory(
+            input_tensor,
+            grid_tensor,
+            output_tensor,
+            padding_mode_,
+            align_corners_,
+            use_precomputed_grid_,
+            batch_output_channels_);
+    }
 }
 
 }  // namespace ttnn::operations::grid_sample
