+++ ttnn/cpp/ttnn/operations/pool/grid_sample/device/grid_sample_op.hpp
@@ -47,11 +47,19 @@ struct GridSample {
 // Program factory declaration
 namespace ttnn::operations::grid_sample {
 
-tt::tt_metal::operation::ProgramWithCallbacks grid_sample_program_factory(
+tt::tt_metal::operation::ProgramWithCallbacks grid_sample_bilinear_program_factory(
     const Tensor& input_tensor,
     const Tensor& grid_tensor,
     const Tensor& output_tensor,
-    const std::string& mode,
+    const std::string& padding_mode,
+    bool align_corners,
+    bool use_precomputed_grid,
+    bool batch_output_channels);
+
+tt::tt_metal::operation::ProgramWithCallbacks grid_sample_nearest_program_factory(
+    const Tensor& input_tensor,
+    const Tensor& grid_tensor,
+    const Tensor& output_tensor,
     const std::string& padding_mode,
     bool align_corners,
     bool use_precomputed_grid,
